bot/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra bot --output-file bot/requirements.txt
anyio==4.6.0
    # via httpx
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
distlib==0.3.8
    # via virtualenv
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
filelock==3.16.1
    # via virtualenv
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via python-telegram-bot
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via pytest
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.8.0
    # via asr-bot (pyproject.toml)
pytest==8.3.3
    # via asr-bot (pyproject.toml)
python-telegram-bot==21.6
    # via asr-bot (pyproject.toml)
pyyaml==6.0.2
    # via pre-commit
requests==2.32.3
    # via asr-bot (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tomli==2.0.2
    # via pytest
typing-extensions==4.12.2
    # via anyio
urllib3==2.2.3
    # via requests
virtualenv==20.26.6
    # via pre-commit
